requirements/requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-databricks

matplotlib
pillow

pytest
pytest-mock

polars
python-dotenv
typing_extensions==4.12.2

databricks-vectorsearch